TypeError: dates.stages[EskomSePushInfo.calc.stage] is not iterable
The wrong schedule was taken (array starts counting at 0, the stage at 1). This
off by one resulted in not being able to find a schedule during stage 8
load shedding. (issue #5)
In case of an event, the start end and time where not stored in the
calculated object, resulting in an "invalid date" status.